+++ src/main/java/com/food/sas/data/dto/OrganizationDTO.java
@@ -4,7 +4,6 @@ import lombok.Data;
 
 import java.io.Serializable;
 import java.util.Date;
-import java.util.List;
 
 /**
  * @author Created by ygdxd_admin at 2019-01-06 8:22 PM
@@ -14,7 +13,7 @@ public class OrganizationDTO implements Serializable {
 
     private static final long serialVersionUID = 1357992997829500968L;
 
-    private Integer id;
+    private Long id;
 
     private String name;
