+++ src/main/java/com/food/sas/data/repository/FocalPictureRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 /**
  * @author Created by ygdxd_admin at 2019-01-08 10:09 PM
  */
-public interface FocalPictureRepository extends MyRepository<FocalPicture, Integer>, QuerydslPredicateExecutor<FocalPicture> {
+public interface FocalPictureRepository extends MyRepository<FocalPicture, Long>, QuerydslPredicateExecutor<FocalPicture> {
 }
