+++ src/main/java/com/food/sas/data/repository/NewsRepository.java
@@ -5,5 +5,5 @@ import com.food.sas.data.entity.News;
 /**
  * @author Created by ygdxd_admin at 2018-12-22 9:14 PM
  */
-public interface NewsRepository extends MyRepository<News, Integer> {
+public interface NewsRepository extends MyRepository<News, Long> {
 }
