+++ src/main/java/com/food/sas/data/repository/OrganizationRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 /**
  * @author Created by ygdxd_admin at 2019-01-06 8:26 PM
  */
-public interface OrganizationRepository extends MyRepository<Organization, Integer>, QuerydslPredicateExecutor<Organization> {
+public interface OrganizationRepository extends MyRepository<Organization, Long>, QuerydslPredicateExecutor<Organization> {
 }
