+++ src/main/java/com/food/sas/data/repository/UserRepository.java
@@ -5,5 +5,5 @@ import com.food.sas.data.entity.User;
 /**
  * @author Created by ygdxd_admin at 2018-12-21 11:05 PM
  */
-public interface UserRepository extends MyRepository<User, Integer> {
+public interface UserRepository extends MyRepository<User, Long> {
 }
